MatchContextを実装
