+++ src/components/molecules/PlayerCard/index.tsx
@@ -4,7 +4,7 @@ type Props = {
   player: Player;
   winCount: number;
   onBreakToggle: (playerId: number) => void;
-  isPlayerInMatch: (playerId: number) => boolean;
+  isPlayerInMatch: boolean;
   isPlayerOnBreak: boolean;
 };
 
@@ -21,7 +21,7 @@ const PlayerCard = ({
       <div>試合数: {player.gamesPlayed}</div>
       <div>勝利数: {winCount}</div>
 
-      {!isPlayerInMatch(player.id) && (
+      {isPlayerInMatch && (
         <div className="ml-auto">
           <button
             className={`bg-blue-500 hover:bg-blue-700 text-white font-bold py-1 px-4 rounded inline-block ${
