+++ src/components/molecules/PlayerStatusCards/index.tsx
@@ -1,23 +1,21 @@
+import { useMatchContext } from "../../../contexts/MatchContext";
 import { usePlayerContext } from "../../../contexts/PlayerContext";
 import PlayerStatusCard from "../../atoms/PlayerStatusCard";
 
-type Props = {
-  isPlayerInMatch: (playerId: number) => boolean;
-};
-
-const PlayerStatusCards: React.FC<Props> = ({ isPlayerInMatch }) => {
+const PlayerStatusCards: React.FC = () => {
   const { players, setOnBreak } = usePlayerContext();
+  const { isPlayerInMatch } = useMatchContext();
   return (
     <>
       {players.map((player) => {
-        const isPlaying = isPlayerInMatch(player.id);
+        const isInMatch = isPlayerInMatch(player.id);
         return (
           <PlayerStatusCard
             key={player.id}
             player={player}
-            isPlaying={isPlaying}
+            isPlaying={isInMatch}
             setOnBreak={setOnBreak}
-            isPlayerInMatch={isPlayerInMatch}
+            isInMatch={isInMatch}
           />
         );
       })}
